Fix bugs

+++ packages/lasotuvi_presentation/lib/src/helper/storage_helper.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:lasotuvi_domain/lasotuvi_domain.dart';
 import 'package:lasotuvi_provider/lasotuvi_provider.dart';
 import 'package:lasotuvi_style/lasotuvi_style.dart';
 import 'package:storage_options/storage_options.dart';
@@ -33,7 +34,12 @@ class StorageHelper {
       onDownload: (uid, item) => onDownload(uid, item, ref),
       onDeleteFromCloud: (uid, item) => onDeleteFromCloud(uid, item, ref),
       onDeleteFromLocal: (item) => onDeleteFromLocal(item, ref),
-      onDeleteForever: (uid, item) => onDeleteForever(uid, item, ref),
+      onDeleteForever: (uid, item) => onDeleteForever(
+        uid: uid,
+        item: item,
+        ref: ref,
+        context: context,
+      ),
       doBeforeDeleteForever: doBeforeDeleteForever,
       callback: callback,
     );
@@ -61,12 +67,23 @@ class StorageHelper {
     );
   }
 
-  static Future<void> onUpload<T>({
+  static Future<bool> onUpload<T>({
     required BuildContext context,
-    required String uid,
+    required String? uid,
     required T item,
     required WidgetRef ref,
   }) async {
+    if (uid == null) {
+      await showDialog(
+        context: context,
+        builder: (_) => const NeedSignInAlertDialog(
+          colorScheme: LasotuviAppStyle.colorScheme,
+          translate: translate,
+        ),
+      );
+      return false;
+    }
+    bool resultOfExcutor = false;
     await ref.read(guardProvider).review().then((value) async {
       if (!value.connected) {
         await showDialog(
@@ -76,6 +93,7 @@ class StorageHelper {
             translate: translate,
           ),
         );
+        resultOfExcutor = false;
       } else if (!value.signedIn) {
         await showDialog(
           context: context,
@@ -84,10 +102,46 @@ class StorageHelper {
             translate: translate,
           ),
         );
+        resultOfExcutor = false;
       } else {
-        await ref.read(uploaderProvider).upload(uid: uid, items: [item]);
+        final allow = await _allow<T>(uid, ref);
+        if (context.mounted) {
+          if (!allow) {
+            await showDialog(
+                context: context,
+                builder: (ctx) => const NeedUpgradePlanDialog(
+                    // goToPlanMarket: () => showDialog(
+                    //       context: ctx,
+                    //       builder: (_) => BasicBottomSheet(
+                    //         title: translate('energyPointMarket'),
+                    //         colorScheme: LasotuviAppStyle.colorScheme,
+                    //         child: const EnergyStoreContainer(),
+                    //       ),
+                    //     ),
+                    translate: translate));
+            resultOfExcutor = false;
+          } else {
+            await ref.read(uploaderProvider).upload(uid: uid, items: [item]);
+            resultOfExcutor = true;
+          }
+        }
       }
     });
+    return resultOfExcutor;
+  }
+
+  static Future<bool> _allow<T>(String uid, WidgetRef ref) async {
+    final currentPlan = await ref.read(takeCurrentStoragePlanProvider)(uid);
+    if (T == Chart) {
+      final count = await ref.read(countChartOnCloudProvider)(uid);
+      return count < currentPlan.limitChart;
+    } else if (T == Tag) {
+      final count = await ref.read(countTagOnCloudProvider)(uid);
+      return count < currentPlan.limitTag;
+    } else {
+      final count = await ref.read(countNoteOnCloudProvider)(uid);
+      return count < currentPlan.limitNote;
+    }
   }
 
   static Future<void> onDownload<T>(String uid, T item, WidgetRef ref) =>
@@ -102,6 +156,20 @@ class StorageHelper {
   ) =>
       ref.read(removerProvider).deleteFromLocal(items: [item]);
 
-  static Future<void> onDeleteForever<T>(String uid, T item, WidgetRef ref) =>
-      ref.read(removerProvider).deleteForever(uid: uid, items: [item]);
+  static Future<void> onDeleteForever<T>({
+    required String uid,
+    required T item,
+    required WidgetRef ref,
+    required BuildContext context,
+  }) async {
+    final result = await showDialog<ConfirmResult>(
+      context: context,
+      builder: (ctx) => const DeleteChartConfirmDialog(translate: translate),
+    );
+    if (result != null) {
+      if (result.yes) {
+        await ref.read(removerProvider).deleteForever(uid: uid, items: [item]);
+      }
+    }
+  }
 }
